Update to upstream version '1.12.0'
with Debian dir af96326e73787f5793b0010944d2968fa9f2e3c9

+++ tests/t-fips-service-ind.c
@@ -1009,8 +1009,8 @@ check_cipher_o_s_e_d_c (int reject)
                     gpg_strerror (err));
 
           if (taglen != tv[tvidx].taglen)
-              fail ("gcry_cipher_info %d failed: taglen mismatch %d != %ld\n", tvidx,
-                    tv[tvidx].taglen, (long)taglen);
+              fail ("gcry_cipher_info %d failed: taglen mismatch %d != %zu\n", tvidx,
+                    tv[tvidx].taglen, taglen);
         }
 
       err = gcry_cipher_encrypt (h, out, MAX_DATA_LEN, pt, blklen);
